requirements.txt
fastapi
uvicorn[standard]
whatap-python
cupy-cuda12x
